app/requirements.txt
alembic
sqlalchemy~=2.0
psycopg2-binary
asyncpg
uvicorn[standard]~=0.17.6
FastApi~=0.78.0
fastapi-camelcase
python-jose
passlib[bcrypt]
python-multipart
pydantic[email]
aiofile
xxhash
